+++ libs/data-service-generator/src/index.ts
@@ -1,4 +1,4 @@
 export * from "@amplication/code-gen-types";
-export { createDataService } from "./create-data-service-impl";
+export { createDataService } from "./create-data-service";
 export { defaultLogger } from "./server/logging";
 export { httpClient } from "./util/http-client";
