+++ libs/data-service-generator/src/tests/create-data-service-with-kafka.spec.ts
@@ -3,7 +3,7 @@ import {
   EnumMessagePatternConnectionOptions,
   Topic,
 } from "@amplication/code-gen-types";
-import { createDataServiceImpl } from "../create-data-service-impl";
+import { createDataService } from "../create-data-service";
 import { EnumResourceType } from "../models";
 import { appInfo, MODULE_EXTENSIONS_TO_SNAPSHOT } from "./appInfo";
 import entities from "./entities";
@@ -65,7 +65,7 @@ describe("createDataService", () => {
         },
       ],
     };
-    const modules = await createDataServiceImpl(service);
+    const modules = await createDataService(service);
     const modulesToSnapshot = modules.filter((module) =>
       MODULE_EXTENSIONS_TO_SNAPSHOT.some((extension) =>
         module.path.endsWith(extension)
